Moved the getFilterClass into its own module
